Update sway script

+++ sway/scripts/screenshot.sh
@@ -8,10 +8,8 @@ while [ -e "${save_path}" ]; do
 
   case "${save_path}" in
   *_*)
-    number="${save_path#*_}"
-    number=$((number + 1))
-    save_path="${save_path%_*}"
-    save_path="Screenshot_${save_path}_${number}.png"
+    number=$((${save_path#*_} + 1))
+    save_path="Screenshot_${save_path%_*}_${number}.png"
     ;;
   *)
     save_path="Screenshot_${save_path}_1.png"
@@ -30,16 +28,18 @@ fi
 
 name="$(tofi -c ~/.config/tofi/dmenu --prompt-text 'Custom filename: ' </dev/null)"
 
-if [ -n "${name}" ]; then
+if [ -n "${name}" ] && [ ! -e "$(dirname "${save_path}")/${name}" ]; then
   case "${name}" in
   */*)
-    notify-send 'Invalid file name.'
+    notify-send 'Invalid filename.' 'Using default filename.'
     ;;
   *)
     mv "${save_path}" "$(dirname "${save_path}")/${name}"
     save_path="$(dirname "${save_path}")/${name}"
     ;;
   esac
+else
+  notify-send 'Invalid filename.' 'Using default filename.'
 fi
 
 answer="$(printf 'No\nYes' | tofi -c ~/.config/tofi/dmenu --prompt-text 'Copy screenshot to clipboard? ')"
